+++ src/main/java/dev/ftb/mods/ftbechoes/client/gui/EchoScreen.java
@@ -22,16 +22,14 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.resources.sounds.SimpleSoundInstance;
-import net.minecraft.client.resources.sounds.SoundInstance;
 import net.minecraft.commands.Commands;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.FormattedText;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.sounds.SoundEvent;
-import net.minecraft.sounds.SoundSource;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.common.util.Lazy;
 import net.neoforged.neoforge.network.PacketDistributor;
 import org.jetbrains.annotations.Nullable;
 
@@ -39,15 +37,14 @@ import java.util.*;
 import java.util.function.BooleanSupplier;
 
 public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
-    static EchoScreen.Page currentPage = EchoScreen.Page.LORE;
+    static Map<ResourceLocation,EchoScreen.Page> currentPage = new HashMap<>();
 
     private final BlockPos projectorPos;
-    private Echo echo;
-    private boolean pendingScrollToEnd;
-    private SoundEvent playingSound = null;
-    private SoundInstance playingSoundInstance = null;
 
-    public EchoScreen(BlockPos projectorPos, Echo echo) {
+    private boolean pendingScrollToEnd;
+    @Nullable private Echo echo;
+
+    public EchoScreen(BlockPos projectorPos, @Nullable Echo echo) {
         super();
 
         this.projectorPos = projectorPos;
@@ -58,6 +55,8 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
     public void setEcho(Echo echo) {
         this.echo = echo;
+        ((TopPanel) topPanel).tabButtons.invalidate();
+        mainPanel.pages.invalidate();
         refreshWidgets();
     }
 
@@ -70,14 +69,16 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
     @Override
     public void onPostInit() {
-        if (currentPage == Page.LORE) {
+        if (getCurrentPage() == Page.LORE) {
             scrollBar.setValue(scrollBar.getMaxValue());
         }
     }
 
     @Override
     public void onClosed() {
-        stopPlayingSound();
+        if (EchoSoundClipHandler.INSTANCE.isPlayingSound() && Minecraft.getInstance().screen != null && Minecraft.getInstance().player != null) {
+            Minecraft.getInstance().player.displayClientMessage(Component.translatable("ftbechoes.message.hold_alt_to_stop_sound").withStyle(ChatFormatting.AQUA), true);
+        }
     }
 
     @Override
@@ -118,9 +119,19 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         }
     }
 
+    public Page getCurrentPage() {
+        return echo == null ? Page.LORE : currentPage.getOrDefault(echo.id(), Page.LORE);
+    }
+
+    private void setCurrentPage(Page page) {
+        if (echo != null) {
+            currentPage.put(echo.id(), page);
+        }
+    }
+
     public void onProgressUpdated() {
         refreshWidgets();
-        if (currentPage == Page.LORE) {
+        if (getCurrentPage() == Page.LORE) {
             pendingScrollToEnd = true;
         }
     }
@@ -129,46 +140,13 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         return projectorPos;
     }
 
-    public void startPlayingSound(SoundEvent soundEvent) {
-        if (playingSoundInstance != null) {
-            Minecraft.getInstance().getSoundManager().stop(playingSoundInstance);
-        }
-        playingSound = soundEvent;
-        playingSoundInstance = createSoundInstance(playingSound);
-        if (playingSoundInstance != null) {
-            Minecraft.getInstance().getSoundManager().play(playingSoundInstance);
-        }
-    }
-
-    public void stopPlayingSound() {
-        startPlayingSound(null);
-    }
-
-    public boolean isPlayingSound(SoundEvent soundEvent) {
-        return playingSound != null && playingSound.getLocation().equals(soundEvent.getLocation());
-    }
-
-    @Nullable
-    private static SimpleSoundInstance createSoundInstance(SoundEvent sound) {
-        if (sound == null) {
-            return null;
-        }
-        return new SimpleSoundInstance(sound.getLocation(), SoundSource.VOICE,
-                1f, 1f, SoundInstance.createUnseededRandom(), false, 0,
-                SoundInstance.Attenuation.NONE,
-                0.0, 0.0, 0.0,
-                true
-        );
-    }
-
-    static class PurchaseButton extends SimpleTextButton {
+    class PurchaseButton extends SimpleTextButton {
         public PurchaseButton(Panel parent) {
             super(parent, Component.translatable("ftbechoes.gui.place_order"), Icons.MONEY_BAG);
         }
 
         @Override
         public void onClicked(MouseButton mouseButton) {
-            // TODO send order to server
             playClickSound();
 
             PacketDistributor.sendToServer(new PlaceOrderMessage(ShoppingBasket.CLIENT_INSTANCE));
@@ -214,13 +192,13 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         @Override
         public boolean isEnabled() {
             return ShoppingBasket.CLIENT_INSTANCE.hasContents()
-                    && currentPage == Page.SHOP
+                    && getCurrentPage() == Page.SHOP
                     && ShoppingBasket.CLIENT_INSTANCE.getTotalCost() <= FTBEchoes.currencyProvider().getTotalCurrency(Minecraft.getInstance().player);
         }
 
         @Override
         public boolean shouldDraw() {
-            return currentPage == Page.SHOP;
+            return getCurrentPage() == Page.SHOP;
         }
     }
 
@@ -228,7 +206,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         private final TextField label;
         private final Button settingsButton;
         private final Button stopAudioButton;
-        private final Map<EchoScreen.Page, EchoScreen.PageButton> buttons = new EnumMap<>(EchoScreen.Page.class);
+        private final Lazy<Map<EchoScreen.Page, EchoScreen.PageButton>> tabButtons = Lazy.of(this::buildTabButtons);
         private final boolean adminPlayer;
 
         public TopPanel() {
@@ -240,14 +218,26 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
             label = new TextField(this);
             settingsButton = SimpleTextButton.create(this, Component.empty(), Icons.SETTINGS, this::showEchoSelector);
             stopAudioButton = new StopAudioButton();
-            buttons.put(EchoScreen.Page.LORE, new EchoScreen.PageButton(EchoScreen.Page.LORE, this, Icons.BOOK));
-            buttons.put(EchoScreen.Page.SHOP, new EchoScreen.PageButton(EchoScreen.Page.SHOP, this, Icons.MONEY_BAG));
+        }
+
+        private Map<EchoScreen.Page, EchoScreen.PageButton> buildTabButtons() {
+            Map<EchoScreen.Page, EchoScreen.PageButton> res = new EnumMap<>(EchoScreen.Page.class);
+
+            res.put(EchoScreen.Page.LORE, new EchoScreen.PageButton(EchoScreen.Page.LORE, this, Icons.BOOK));
+            if (echo != null && echo.hasAnyShopItems()) {
+                res.put(EchoScreen.Page.SHOP, new EchoScreen.PageButton(EchoScreen.Page.SHOP, this, Icons.MONEY_BAG));
+            }
 
             // only show selector drop-down after at least one stage has been completed
-            buttons.get(Page.LORE).setDropdownAction(
+            res.get(Page.LORE).setDropdownAction(
                     this::showStageSelector,
                     () -> echo != null && ClientProgress.get().isStageCompleted(echo.id(), 0)
             );
+            return res;
+        }
+
+        private Map<EchoScreen.Page, EchoScreen.PageButton> getTabButtons() {
+            return tabButtons.get();
         }
 
         @Override
@@ -261,9 +251,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
             } else {
                 add(label.setText(Component.empty().withStyle(ChatFormatting.YELLOW).append(echo.title())));
             }
-            for (EchoScreen.Page p : buttons.keySet()) {
-                add(buttons.get(p));
-            }
+            addAll(getTabButtons().values());
         }
 
         @Override
@@ -274,9 +262,9 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
             label.setPos(4, 5);
             label.setWidth(width);
 
-            int bw = width / (buttons.size() + 1) - 4;
+            int bw = width / (getTabButtons().size() + 1) - 4;
             int by = getTheme().getFontHeight() + 8;
-            for (PageButton w : buttons.values()) {
+            for (EchoScreen.PageButton w : getTabButtons().values()) {
                 w.setPosAndSize(4 + w.page.ordinal() * (bw + 2), by, bw, height - by);
             }
         }
@@ -285,9 +273,13 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         public void drawBackground(GuiGraphics graphics, Theme theme, int x, int y, int w, int h) {
             super.drawBackground(graphics, theme, x, y, w, h);
             int col = 0xff585d66;  // blends best with tabbed outline
-            graphics.hLine(x, x + buttons.get(Page.LORE).posX - 1, y + height - 2, col);
-            graphics.hLine(x + buttons.get(Page.LORE).posX + buttons.get(Page.LORE).width, x + buttons.get(Page.SHOP).posX - 1, y + height - 2, col);
-            graphics.hLine(x + buttons.get(Page.SHOP).posX + buttons.get(Page.SHOP).width, x + width - 1, y + height - 2, col);
+            graphics.hLine(x, x + getTabButtons().get(Page.LORE).posX - 1, y + height - 2, col);
+            if (echo != null && echo.hasAnyShopItems()) {
+                graphics.hLine(x + getTabButtons().get(Page.LORE).posX + getTabButtons().get(Page.LORE).width, x + getTabButtons().get(Page.SHOP).posX - 1, y + height - 2, col);
+                graphics.hLine(x + getTabButtons().get(Page.SHOP).posX + getTabButtons().get(Page.SHOP).width, x + width - 1, y + height - 2, col);
+            } else {
+                graphics.hLine(x + getTabButtons().get(Page.LORE).posX + getTabButtons().get(Page.LORE).width, x + width - 1, y + height - 2, col);
+            }
             graphics.hLine(x, x + width - 1, y + height - 1, NordColors.POLAR_NIGHT_1.rgba());
         }
 
@@ -314,6 +306,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         }
 
         private void showStageSelector() {
+            if (echo == null) return;
             openContextMenu(Util.make(new ArrayList<>(), list -> {
                         list.add(ContextMenuItem.title(Component.translatable("ftbechoes.gui.stages")));
                         list.add(ContextMenuItem.separator());
@@ -341,7 +334,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         }
 
         private void scrollToStage(int stageIdx) {
-            if (EchoScreen.this.mainPanel.pages.get(Page.LORE) instanceof LorePanel lorePanel) {
+            if (EchoScreen.this.mainPanel.getPages().get(Page.LORE) instanceof LorePanel lorePanel) {
                 if (lorePanel.isCollapsed(stageIdx)) {
                     lorePanel.setCollapsed(stageIdx, false);
                 }
@@ -351,7 +344,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
         private void collapseAll(boolean collapse) {
             playClickSound();
-            if (EchoScreen.this.mainPanel.pages.get(Page.LORE) instanceof LorePanel lorePanel) {
+            if (EchoScreen.this.mainPanel.getPages().get(Page.LORE) instanceof LorePanel lorePanel) {
                 lorePanel.setAllCollapsed(collapse);
             }
         }
@@ -369,39 +362,56 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
             @Override
             public void onClicked(MouseButton mouseButton) {
-                EchoScreen.this.stopPlayingSound();
+                EchoSoundClipHandler.INSTANCE.stopPlayingSound();
             }
 
             @Override
             public boolean shouldDraw() {
-                return EchoScreen.this.playingSound != null;
+                return isEnabled();
+            }
+
+            @Override
+            public boolean isEnabled() {
+                return EchoSoundClipHandler.INSTANCE.isPlayingSound();
             }
 
             @Override
             public void addMouseOverText(TooltipList list) {
-                list.add(Component.translatable("ftbechoes.gui.stop_audio"));
+                if (shouldDraw()) {
+                    list.add(Component.translatable("ftbechoes.gui.stop_audio"));
+                }
             }
         }
     }
 
     public class MainPanel extends Panel {
-        private final Map<Page, PagePanel> pages = new EnumMap<>(Page.class);
+        private final Lazy<Map<Page, PagePanel>> pages = Lazy.of(this::buildPages);
 
         public MainPanel() {
             super(EchoScreen.this);
+        }
 
-            pages.put(Page.LORE, new LorePanel(this, EchoScreen.this));
-            pages.put(Page.SHOP, new ShopPanel(this, EchoScreen.this));
+        private Map<Page, PagePanel> buildPages() {
+            Map<Page,PagePanel> map = new EnumMap<>(Page.class);
+            map.put(Page.LORE, new LorePanel(this, EchoScreen.this));
+            if (echo != null && echo.hasAnyShopItems()) {
+                map.put(Page.SHOP, new ShopPanel(this, EchoScreen.this));
+            }
+            return map;
+        }
+
+        private Map<Page,PagePanel> getPages() {
+            return pages.get();
         }
 
         @Override
         public void addWidgets() {
-            addAll(pages.values());
+            addAll(getPages().values());
         }
 
         @Override
         public void alignWidgets() {
-            pages.values().forEach(p -> {
+            getPages().values().forEach(p -> {
                 p.setPosAndSize(0, 0, width, height);
                 p.alignWidgets();
             });
@@ -410,21 +420,30 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
             }
         }
 
+        private PagePanel getCurrentPagePanel() {
+            PagePanel res = getPages().get(getCurrentPage());
+            if (res == null) {
+                setCurrentPage(Page.LORE);
+                return getPages().get(Page.LORE);
+            }
+            return res;
+        }
+
         @Override
         public int getContentHeight() {
-            return pages.get(currentPage).getContentHeight();
+            return getCurrentPagePanel().getContentHeight();
         }
 
         public void storeScrollPos() {
-            pages.get(currentPage).storeScrollPos();
+            getCurrentPagePanel().storeScrollPos();
         }
 
         public double fetchScrollPos() {
-            return pages.get(currentPage).fetchScrollPos();
+            return getCurrentPagePanel().fetchScrollPos();
         }
 
         public void onSwitchAway() {
-            pages.get(currentPage).onSwitchAway();
+            getCurrentPagePanel().onSwitchAway();
         }
 
         @Override
@@ -433,7 +452,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
         }
     }
 
-    private static class BottomPanel extends Panel {
+    private class BottomPanel extends Panel {
         private final PurchaseButton purchaseButton;
 
         public BottomPanel(Panel panel) {
@@ -482,12 +501,12 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
         @Override
         public boolean shouldDraw() {
-            return page == currentPage;
+            return page == echoScreen.getCurrentPage();
         }
 
         @Override
         public boolean isEnabled() {
-            return page == currentPage;
+            return page == echoScreen.getCurrentPage();
         }
 
         protected Optional<Echo> getEcho() {
@@ -504,10 +523,6 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
         public double fetchScrollPos() {
             return lastScrollPos;
-        }
-
-        public void scrollToTop() {
-            echoScreen.scrollBar.setValue(0.0);
         }
 
         public void onSwitchAway() {
@@ -580,12 +595,12 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
         @Override
         public void drawBackground(GuiGraphics graphics, Theme theme, int x, int y, int w, int h) {
-            theme.drawHorizontalTab(graphics, x, y, w, h, EchoScreen.currentPage == page);
+            theme.drawHorizontalTab(graphics, x, y, w, h, getCurrentPage() == page);
         }
 
         @Override
         public void onClicked(MouseButton mouseButton) {
-            if (dropdownPredicate.getAsBoolean() && EchoScreen.currentPage == page && getMouseX() < getX() + width - 4 && getMouseX() > getX() + width - 18) {
+            if (dropdownPredicate.getAsBoolean() && getCurrentPage() == page && getMouseX() < getX() + width - 4 && getMouseX() > getX() + width - 18) {
                 onDropDownClicked.run();
                 return;
             }
@@ -595,7 +610,7 @@ public class EchoScreen extends AbstractThreePanelScreen<EchoScreen.MainPanel> {
 
             mainPanel.onSwitchAway();
             mainPanel.storeScrollPos();
-            EchoScreen.currentPage = page;
+            setCurrentPage(page);
             mainPanel.refreshWidgets();
             scrollBar.setValue(mainPanel.fetchScrollPos());
         }