+++ src/main/java/dev/ftb/mods/ftbechoes/datagen/ModSoundProvider.java
@@ -17,5 +17,12 @@ public class ModSoundProvider extends SoundDefinitionsProvider {
         add(ModSounds.COINS, SoundDefinition.definition()
                 .with(sound(FTBEchoes.id("coins")))
                 .subtitle("ftbechoes.subtitle.coins_clink"));
+
+        add(ModSounds.THIS_IS_AN_ECHO, SoundDefinition.definition()
+                .with(sound(FTBEchoes.id("this_is_an_echo"))));
+        add(ModSounds.THIS_IS_AN_ECHO_EN_GB, SoundDefinition.definition()
+                .with(sound(FTBEchoes.id("en_gb/this_is_an_echo"))));
+        add(ModSounds.THIS_IS_AN_ECHO_FR_FR, SoundDefinition.definition()
+                .with(sound(FTBEchoes.id("fr_fr/this_is_an_echo"))));
     }
 }